Move settings forms to module.
